plus initial test
More tests.

I've broken these tests out into their own file. An item of
followup work this presents is to move the helper function
and using a constant for a url prefix pattern to the remainder
of the tests as well, but I feel that is beyond the scope
of this already large issue.

friendlier imports for non2.7

2.6 hates format without numbers