rm pycryptodome
Scoped Keys are deprecated and pycryptodome is not pure python
making it a problem for people wanting to `pip install keen` on
some systems. pycryptodome is a dependency of only scoped keys.

To make the tests pass, you'll need to have it installed but that
is enabled via the test_requires within setup.py. A fun error
message is displayed when anything scoped keys are imported. This
may make `from keen import *` fail, but that's bad python to begin
with so I don't feel very bad about that.

This includes a documentation update reflecting the manual step
required for the deprecated feature.